% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AlignRefAndRead}
\alias{AlignRefAndRead}
\title{Align a read and a reference window.  Both the forward and reverse direction
of the read will be  aligned, and only the highest scoring alignment will be
returned.}
\usage{
AlignRefAndRead(ref, read)
}
\arguments{
\item{ref}{A string that is the reference window we are aligning to.}

\item{read}{A list with elements name/read.}
}
\value{
Returns a list with the aligned read, ref and score
}
\description{
Align a read and a reference window.  Both the forward and reverse direction
of the read will be  aligned, and only the highest scoring alignment will be
returned.
}

