% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AlnsToDataFrame}
\alias{AlnsToDataFrame}
\title{Takes a list of pairwise alignments and generates a data frame with the MSA
sequence. The reference sequence is determined from the first alignment in
the list. All reads are expanded by adding enough inserts to account for all
of them.}
\usage{
AlnsToDataFrame(alns)
}
\arguments{
\item{ref}{A string that is the reference window we are aligning to.}

\item{read}{A list with elements name/read.}
}
\value{
Returns a list with the aligned read, ref and score
}
\description{
Takes a list of pairwise alignments and generates a data frame with the MSA
sequence. The reference sequence is determined from the first alignment in
the list. All reads are expanded by adding enough inserts to account for all
of them.
}

