% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\name{getRefPosition}
\alias{getRefPosition}
\title{Determine the reference position in a multiple sequence alignment returned
from AlnsToDataFrame.}
\usage{
getRefPosition(tstart, pos, df)
}
\arguments{
\item{tstart}{The initial position of the alignment.}

\item{pos}{The}

\item{df}{Data frame returned by AlnsToDataFrame.  The seq in the}

\item{fasta_name}{Full path of the fasta file used to generate the CCS alignment}
}
\value{
Returns The genome position
}
\description{
Typically, one will call plotMSA to view a variant call in an alignment, but
that function takes coordinates in terms of alignment positions, not
reference genome positions.  This function converts a reference genome
location to an alignment location.  It does this by taking the start of the
alignment on the genome, and then figures out how much further it needs to go
to get the respective genome position in the alignment.
}

