% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotMSA}
\alias{plotMSA}
\title{Print a PDF of the MSA in a dataframe returned by the AlnsToDataFrame function.}
\usage{
plotMSA(df, pdfname, start = 0, end = -1, showPositions = TRUE,
  useOrigColors = FALSE)
}
\arguments{
\item{df}{A data frame returned by a call to AlnsToDataFrame.}

\item{pdfname}{The name of the pdf file to output}

\item{start}{Optionally subset a portion of the alignment by specifying a start column.}

\item{end}{Optionally subset a portion of the alignment by specifying an end column.}

\item{useOrigColors}{Whether to use my original color scheme http://colorbrewer2.org/?type=qualitative&scheme=Set1&n=4}
}
\value{
Returns a grid object with the alignment
}
\description{
Print a PDF of the MSA in a dataframe returned by the AlnsToDataFrame function.
}

